Made video annotations a separate fixture
